Add a separate project view-model class
The project view-model is a proxy between the views and controllers of
the of the document UI, on one side, and the Haskell-side model of the
project as a Cabal package, on the other side.

The view-model owns all objects represented, but not owned by the views
(such as the items of the outline view).