Truncate overly long interpreter results
The problem with #272 was the unbounded result string. Even very long
strings were leading to hang in the UI (for mysterious reasons to do
with how the Cocoa event source gets drained). We simply use a fixed
cutoff.

There is no point in making the cutoff customisable as it has too much
of a whole system performance impact when the string gets too long.

Fixes #272