Disable responsive scrolling for code views
During responsive scrolling, `visibleRect` lags behind the actual
visible portion of the scrolled view (as the scrolling happens on a
background thread, but `visibleRect` must be updated on the main
thread). Hence, we switch of responsive scrolling for code views for
now. Otherwise, the gutter view and `SynchoScrollView` lag during
scrolling.

See also WWDC 2013, Session 215: ”Optimizing Drawing and Scrolling on
OS X”.